Clean up engagements refactor
