+++ app-modules/engagement/src/Filament/Actions/MessageCenterSendEngagementAction.php
@@ -43,7 +43,6 @@ use AdvisingApp\Engagement\Models\EmailTemplate;
 use AdvisingApp\Engagement\Models\Engagement;
 use AdvisingApp\Notification\Enums\NotificationChannel;
 use AdvisingApp\StudentDataModel\Models\Contracts\Educatable;
-use App\Features\EngagementsFeature;
 use Filament\Actions\Action;
 use Filament\Actions\StaticAction;
 use Filament\Forms\Components\Actions;
@@ -169,7 +168,7 @@ class MessageCenterSendEngagementAction extends Action
                         Toggle::make('send_later')
                             ->reactive()
                             ->helperText('By default, this email or text will send as soon as it is created unless you schedule it to send later.'),
-                        DateTimePicker::make('deliver_at')
+                        DateTimePicker::make('scheduled_at')
                             ->required()
                             ->visible(fn (Get $get) => $get('send_later')),
                     ]),
@@ -184,27 +183,21 @@ class MessageCenterSendEngagementAction extends Action
                     $this->halt();
                 }
 
-                if (EngagementsFeature::active()) {
-                    $engagement = app(CreateEngagement::class)->execute(new EngagementCreationData(
-                        user: auth()->user(),
-                        recipient: $this->getEducatable(),
-                        channel: NotificationChannel::parse($data['channel']),
-                        subject: $data['subject'] ?? null,
-                        body: $data['body'] ?? null,
-                        temporaryBodyImages: array_map(
-                            fn (TemporaryUploadedFile $file): array => [
-                                'extension' => $file->getClientOriginalExtension(),
-                                'path' => (fn () => $this->path)->call($file),
-                            ],
-                            $form->getFlatFields()['body']->getTemporaryImages(),
-                        ),
-                        scheduledAt: ($data['send_later'] ?? false) ? Carbon::parse($data['deliver_at'] ?? null) : null,
-                    ));
-                } else {
-                    $engagement = new Engagement($data);
-                    $engagement->recipient()->associate($this->getEducatable());
-                    $engagement->save();
-                }
+                $engagement = app(CreateEngagement::class)->execute(new EngagementCreationData(
+                    user: auth()->user(),
+                    recipient: $this->getEducatable(),
+                    channel: NotificationChannel::parse($data['channel']),
+                    subject: $data['subject'] ?? null,
+                    body: $data['body'] ?? null,
+                    temporaryBodyImages: array_map(
+                        fn (TemporaryUploadedFile $file): array => [
+                            'extension' => $file->getClientOriginalExtension(),
+                            'path' => (fn () => $this->path)->call($file),
+                        ],
+                        $form->getFlatFields()['body']->getTemporaryImages(),
+                    ),
+                    scheduledAt: ($data['send_later'] ?? false) ? Carbon::parse($data['scheduled_at'] ?? null) : null,
+                ));
 
                 $form->model($engagement)->saveRelationships();
             })
