+++ app-modules/timeline/src/Timelines/EngagementTimeline.php
@@ -40,7 +40,6 @@ use AdvisingApp\Engagement\Models\Engagement;
 use AdvisingApp\Notification\Enums\NotificationChannel;
 use AdvisingApp\Notification\Enums\NotificationDeliveryStatus;
 use AdvisingApp\Timeline\Models\CustomTimeline;
-use App\Features\EngagementsFeature;
 use Filament\Actions\ViewAction;
 use Filament\Infolists\Components\Fieldset;
 use Filament\Infolists\Components\IconEntry;
@@ -65,11 +64,7 @@ class EngagementTimeline extends CustomTimeline
 
     public function sortableBy(): string
     {
-        if (EngagementsFeature::active()) {
-            return $this->engagement->scheduled_at ?? $this->engagement->created_at;
-        }
-
-        return $this->engagement->deliver_at;
+        return $this->engagement->scheduled_at ?? $this->engagement->created_at;
     }
 
     public function providesCustomView(): bool
